Refactor/features/renaming features (#368)
