+++ eegdash/features/__init__.py
@@ -23,7 +23,9 @@ from .extractors import FeatureExtractor
 from .feature_bank import (  # Complexity; Connectivity; CSP; Dimensionality; Signal; Spectral
     CommonSpatialPattern,
     complexity_approx_entropy,
+    complexity_detrended_fluctuation_analysis,
     complexity_entropy_preprocessor,
+    complexity_hurst_exp,
     complexity_lempel_ziv,
     complexity_multiscale_entropy,
     complexity_sample_entropy,
@@ -41,17 +43,15 @@ from .feature_bank import (  # Complexity; Connectivity; CSP; Dimensionality; Si
     connectivity_spectral_correlation,
     connectivity_temporal_correlation,
     connectivity_weighted_phase_lag_index,
-    dimensionality_detrended_fluctuation_analysis,
     dimensionality_higuchi_fractal_dim,
-    dimensionality_hurst_exp,
     dimensionality_katz_fractal_dim,
     dimensionality_petrosian_fractal_dim,
     pick_channel_pairs_preprocessor,
     pick_channels_preprocessor,
     preprocessor_as_feature,
+    signal_bandpass_filter_preprocessor,
     signal_decorrelation_time,
     signal_envelope_preprocessor,
-    signal_filter_preprocessor,
     signal_hjorth_activity,
     signal_hjorth_complexity,
     signal_hjorth_mobility,
@@ -150,6 +150,8 @@ __all__ = [
     "complexity_sample_entropy",
     "complexity_svd_entropy",
     "complexity_lempel_ziv",
+    "complexity_hurst_exp",
+    "complexity_detrended_fluctuation_analysis",
     # Connectivity
     "connectivity_coherency_preprocessor",
     "connectivity_phase_diff_preprocessor",
@@ -170,13 +172,11 @@ __all__ = [
     "dimensionality_higuchi_fractal_dim",
     "dimensionality_petrosian_fractal_dim",
     "dimensionality_katz_fractal_dim",
-    "dimensionality_hurst_exp",
-    "dimensionality_detrended_fluctuation_analysis",
     # Pick
     "pick_channel_pairs_preprocessor",
     "pick_channels_preprocessor",
     # Signal
-    "signal_filter_preprocessor",
+    "signal_bandpass_filter_preprocessor",
     "signal_envelope_preprocessor",
     "signal_mean",
     "signal_variance",
